flaskapp/app/craft/requirements.txt
cycler==0.10.0
astor==0.8.1
certifi==2019.9.11
gast==0.2.2
google-pasta==0.1.7
grpcio==1.27.1
h5py==2.9.0
idna==2.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
numpy==1.18.1
opt-einsum==3.0.1
opencv-python==4.1.2.30
pipenv==2018.11.26
protobuf==3.9.1
python-dateutil==2.8.0
pyxdg==0.26
PyYAML==5.1.2
six==1.11.0
tb-nightly==1.15.0a20190806
tensorboard==2.0.0
tensorflow-estimator==2.0.0
tensorflow-gpu==2.0.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019080601
tqdm==4.42.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
Werkzeug==0.15.5
wrapt==1.11.2
keras-ocr==0.6.3
imgaug==0.4.0
sklearn